Added Documentation
Added `docs/PROJECT_PHASES.md` to outline development phases.

Created a `README.md` for the `docs/` better navigation and understand of the documentation.
Created `getting-started.md and troubleshooting.md ` files for more clarity on local and Docker setup. for clarity on local and Docker setup.
Added notes to api.md for the flow of logic in the entry-point file.