new feat to toggle completed for a todo item and refactored the response handling
`App.jsx`:
refactored the handling of the response object in the <App/> by declaring `responseJSON to parse JSON into an object and then access its properties on the following line.

New feat logic, for showing a todo item as `completed`, is added in the `toggleCompleted` callback. A `PUT` request sent to the server once called, therefore add controller function and routing path in `todosController.js` line 69 and `routes/todos.js` line 10, respectively.

updated the names of the access functions and controller functions in `data/todoStore.js` and `todosController.js`, respectively

implemented the new property for the toggle completion, in the in-memory storage: line 8 and 13. Used this property in the toggleCompletedData() to flip the boolean value, upon user click event.

`onToggleCompleted` prop passed into the `TodoItem.jsx` and `TodoList.jsx` files to be called during the user click event on these component. `input type=checkbox` element added to the `<li/>` item element in Todo.Item.jsx for rendering the checkbox.