refactor: added atomic component for rendering <Button>
dynamically created a consistent button component with styling, to better manage the styling of the frontend during development. in `components/common/button.jsx`
a switch statement was used to conditionally render a specific style of <Button> based on the custom `variant` prop, in its parent components. `line 21-36`
This component is imported into the other child components in `src/component` dir, and the `src/App.jsx` file.