_tests_: Started unit testing on the business logic level
Initialised the root directory, to install and run tests using `jest` package. `./package.json` created and `./node_modules` and `package-lock.json` file added to `.gitignore`.

testing 2/3 of the user story scenarios in `add-todo.feature`, in `add-todo.test.js`. This was done by verifying the error validation of the controller functions created on the business logic level of the application. Mock response objects had to be created in order to call the controllers in Jest.