tests: Edit-todo testing on client-side
added user story in `edit-todo.feature` and created test cases on these scenarios in the `edit-todo-integration.test.jsx` file.

`edit-todo-integration.test.jsx`:

Additional callback functions was added in order to manage the piece of state in the test environment. `line 17-30`

In order to complete the test case for a bulk update(`line 134-150`), a refactor was needed to the todo list in the piece of state: to hold more than one todo item `todos` on `line 13`.

on `line 95-101`, user event  added to close the notification component after the test case.

additional test cases were created (`line 104-130`) for validating the edit button and checkbox functionality, on each todo item.